Add update_mentioned_people/2
