+++ lib/operately/operations/comment_adding/subscriptions.ex
@@ -1,6 +1,5 @@
 defmodule Operately.Operations.CommentAdding.Subscriptions do
   alias Ecto.Multi
-  alias Operately.{Notifications, RichContent}
   alias Operately.Notifications.SubscriptionList
 
   def update(multi, :project_check_in_commented, content), do: execute_update(multi, content)
@@ -10,7 +9,7 @@ defmodule Operately.Operations.CommentAdding.Subscriptions do
   defp execute_update(multi, content) do
     multi
     |> fetch_subscriptions()
-    |> update_subscriptions(content)
+    |> Operately.Operations.Notifications.Subscription.update_mentioned_people(content)
   end
 
   defp fetch_subscriptions(multi) do
@@ -20,33 +19,5 @@ defmodule Operately.Operations.CommentAdding.Subscriptions do
         preload: :subscriptions
       ])
     end)
-  end
-
-  defp update_subscriptions(multi, content) do
-    ids = RichContent.find_mentioned_ids(content, :decode_ids)
-
-    Enum.reduce(ids, multi, fn id, multi ->
-      name = "subscription_" <> id
-
-      Multi.run(multi, name, fn _, changes ->
-        if subscription_exists?(changes, id) do
-          {:ok, nil}
-        else
-          Notifications.create_subscription(%{
-            subscription_list_id: changes.subscription_list.id,
-            person_id: id,
-            type: :mentioned,
-          })
-        end
-      end)
-    end)
-  end
-
-  #
-  # Helpers
-  #
-
-  defp subscription_exists?(changes, id) do
-    Enum.any?(changes.subscription_list.subscriptions, &(&1.person_id == id))
   end
 end
